==> code/05 merge-validation_partial.qmd <==
---
title: "Merging layers and validation"
format: html
---

# Learning Objectives  
Today's objectives are to:  
- **Import** the processed (cleaned and interpolated) layers

- Spatially **join** the layers above into one object  

- Validate the spatial-temporal stability zones  

# Setup  
```{r }
# Packages
library(dplyr)
library(tidyr)
library(readr)
library(sf) #vector manipulation
library(ggplot2)
library(viridis)
library(ggthemes)
library(patchwork) #combining multiple plots
library(forcats)
```

```{r defining map theme}
mytheme_map <- 
  theme_map()+
  theme(legend.position = "right",
        panel.background = element_rect(fill="gray80",
                                        color=NA),
        plot.background = element_rect(fill="gray80",
                                       color=NA),
        legend.title = element_text(size=30),
        legend.text = element_text(size=30),
        plot.title = element_text(hjust = 0.5, colour = "blue",size = 32),
        plot.caption = element_text(hjust = 0.5, colour = "blue", size = 28)
        )

```

# Data import  
Let's import the layers we have produced until this point, including:  
- Spatial-temporal yield stability zones    

- Terrain variables (elevation, slope, aspect, flow direction)  

- Soil ECa (0-30 and 0-90 cm layers)  

```{r sty_v}
sy16_v <- read_sf("../data/sy_16_yield_class.geojson")

sy16_v
```

```{r terrain_v}
terrain_v <- read_sf("../data/terrain_v.geojson")

terrain_v
```

```{r eca_v }
eca_v <- read_sf("../data/eca_v.geojson")

eca_v
```
```{r boundary_w}
boundary_w0 <- read_sf("../data/boundary/DR1 boundary.shp")%>%
  st_transform(crs=6345)

boundary_w <- boundary_w0[2,]
boundary_w
```
```{r} 
all_v <- terrain_v %>%
  st_join(eca_v, join = st_equals,
          left = T
          )%>%
  relocate(geometry, .after = last_col()) # to relocate the column
all_v
```

# Joining all data together  donot run
```{r all_v}
all_v <- sy16_v %>%
  st_join(terrain_v, join = st_equals,
          left = T
          ) %>%
  st_join(eca_v, join = st_equals,
          left = T
          )%>%
  relocate(geometry, .after = last_col()) # to relocate the column
all_v
```

# Exporting  
```{r}
write_sf(all_v, "../data/all_v.geojson", delete_dsn = T)
```

# EDA  
```{r summary}
summary(all_v)
```
Where do we see **NAs**? Why is that?  

```{r sty_map}
sty_map <-ggplot()+
  geom_sf(data = all_v, aes(fill = yield_class), color = NA) +
  scale_fill_viridis_d()+
  geom_sf(data = boundary_w, fill = NA)+
  mytheme_map

sty_map
```
```{r elev_map}
elev_map <- ggplot()+
  geom_sf(data = all_v, 
          aes(fill = elev_m),
          color = NA
  )+
  scale_fill_viridis_b()+
  geom_sf(data = boundary_w, fill = NA)+
  mytheme_map

elev_map
```

```{r slope_map}
slope_map <- ggplot()+
  geom_sf(data = all_v, 
          aes(fill = slope),
          color = NA
  )+
  scale_fill_viridis_b()+
  geom_sf(data = boundary_w, fill = NA)+
  mytheme_map

slope_map
```
```{r aspect_map}
aspect_map <- ggplot()+
  geom_sf(data = all_v, 
          aes(fill = aspect),
          color = NA
  )+
  scale_fill_viridis_c()+
  geom_sf(data = boundary_w, fill = NA)+
  mytheme_map

aspect_map
```

```{r eca30_map}
eca30_map <- ggplot()+
  geom_sf(data = all_v, 
          aes(fill = eca30_dsm),
          color = NA
  )+
  scale_fill_viridis_b()+
  geom_sf(data = boundary_w, fill = NA)+
  mytheme_map

eca30_map
```
```{r eca90_map}
eca90_map <- ggplot()+
  geom_sf(data = all_v, 
          aes(fill = eca90_dsm),
          color = NA
  )+
  scale_fill_viridis_b()+
  geom_sf(data = boundary_w, fill = NA)+
  mytheme_map

eca90_map
```
```{r all maps}
#(sty_map + 
   (eca30_map + eca90_map) /
  (elev_map + slope_map + aspect_map)

ggsave("../output/allmaps.png",
       width = 10,
       height = 5
)
```

Just by looking at the maps, does it seem like there may be variables correlated with stability zones?  


# Validation  
Now let's validate the stability zones with the terrain and ec data layers.  

For that, let's start exploring using boxplots.  

```{r boxplots}
all_v %>%
  dplyr::select(-flowdir) %>%
  pivot_longer(cols = elev_m:eca90_dsm) %>%
  mutate(yield_class=factor(yield_class,
                        levels = c("low-yield","medium-yield","high-yield")))%>%
   mutate(name=factor(name,
                        levels = c("elev_m","slope","aspect","eca30_dsm", "eca90_dsm")))%>%
  ggplot(aes(x = yield_class, y = value))+
  geom_boxplot(aes(fill=yield_class))+
  scale_fill_viridis_d(option = "C")+
  facet_wrap(~name, scales = "free_y")+
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "none", axis.title = element_blank()
        )
  

```

Visually, what can we infer?  

- For elevation, which stability class seems to have greater values?  
high-stable

- What about for slope?  
medium-stable
- What about aspect?  
high-stable
- What about for eca shallow?
medium-stable
- What about for eca deep?
high-stable

These are just visual hypothesis.  

How can we statistically assess if the boxplots are different across stability classes?  

Using **Analysis of variance (ANOVA)**.

We are not going to learn this step-by-step as that is beyond the scope of this class. However, you will find below code that, iteratively for each field variable (elevation, ec, etc.):    

- runs ANOVA explained by the different stability classes  
- extracts ANOVA table to check significance  
- extracts ANOVA model means  
- performs and extracts pairwise comparisons with letter separation  



```{r ANOVAs}
#install.packages("emmeans")
#install.packages("multcomp")
#install.packages("multcompView")  #multi comparasion
#install.packages("purrr")   # can do iteration
#install.packages("car")

library(emmeans)
library(multcomp)
library(purrr)
library(car)# for anova

anovas <- all_v %>%
  dplyr::select(-flowdir) %>%
  pivot_longer(cols = elev_m:eca90_dsm) %>%
  group_by(name) %>%
  nest() %>%    #nest groups and nested 
  mutate(mod = map(data,
                   ~lm(data = .x,
                       formula = value ~ yield_class
                   )
  )) %>%
  mutate(anova = map(mod,
                     ~Anova(.x))) %>%
  mutate(cld = map(mod,
                   ~emmeans(.x, ~yield_class) %>%
                     cld(reversed = T, 
                         alpha = .05,
                         Letter = letters) %>%
                     as.data.frame() %>%
                     mutate(letter = trimws(.group))
  )) %>%
  unnest(cld) %>%
  ungroup() %>%
  mutate(name = factor(name,
                       levels = c("elev_m",
                                  "slope",
                                  "aspect",
                                  "eca30_dsm",
                                  "eca90_dsm")))


anovas

```

```{r boxplots and letters}
all_v %>%
  dplyr::select(-flowdir) %>%
  pivot_longer(cols = elev_m:eca90_dsm) %>%
  mutate(yield_class=factor(yield_class,
                        levels = c("low-yield","medium-yield","high-yield")))%>%
   mutate(name=factor(name,
                        levels = c("elev_m","slope","aspect","eca30_dsm", "eca90_dsm")))%>%
  ggplot(aes(x = yield_class, y = value))+
  geom_boxplot(aes(fill=yield_class))+
  geom_label(data = anovas,
             aes(y = emmean,
                 label = letter
                 )
             )+
  scale_fill_viridis_d(option = "C")+
  facet_wrap(~name, scales = "free_y")+
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position = "none", axis.title = element_blank()
        )
```

High-stable areas have:  
- elevation: highest
- slope: lowest
- aspect: highest
- eca shallow: intermediate
- eca deep: highest

Low-stable areas have:
- elevation:  lowest 
- slope: intermediate
- aspect: lowest
- eca shallow: lowest
- eca deep: lowest

Unstable areas have:  
- elevation: lowest
- eca shallow:lowest
- eca deep: lowest
- slope:  lowest
- aspect: lowest


# Quiz 
- Based on the above, explain why high-stable areas are such and your hypothesis of why that is.   


- Based on the above, explain why low-stable areas are such and your hypothesis of why that is.  


- Based on the above, explain why unstable areas are such and your hypothesis of why that is (relate your answer to how different weather patterns can make it high-yielding some years and low-yielding in others as weather interacts with the variables above). 

Send your answers to **lmbastos@uga.edu**, this will also count as your attendance.  

# Summary  
Today we:  
- Brought in all processed layers, from yield to eca.  
- Joined these in one object and exported it for future use  
- Validated stability zones  

# Next steps  
Next, we will create management zones based solely on terrain and ec variables.
